one web desktop per screen

+++ CustomWindow.m
@@ -2,13 +2,12 @@
 
 @implementation CustomWindow
 
-- (id)initWithContentRect:(NSRect)contentRect styleMask:(unsigned int)aStyle backing:(NSBackingStoreType)bufferingType defer:(BOOL)flag 
-{
-    NSWindow* result = [super initWithContentRect:contentRect styleMask:aStyle backing:bufferingType defer:flag];
+- (id)initWithContentRect:(NSRect)contentRect styleMask:(NSUInteger)aStyle backing:(NSBackingStoreType)bufferingType defer:(BOOL)flag {
+    NSWindow *result = [super initWithContentRect:contentRect styleMask:aStyle backing:bufferingType defer:flag];
 
     [result setBackgroundColor:[NSColor clearColor]];
     [result setOpaque:NO];
-	
+
     return result;
 }
 
@@ -16,14 +15,12 @@
 // Custom windows that use the NSBorderlessWindowMask can't become key by default.  Therefore, controls in such windows
 // won't ever be enabled by default.  Thus, we override this method to change that.
 
-- (BOOL)canBecomeKeyWindow
-{
+- (BOOL)canBecomeKeyWindow {
     return YES;
 }
 
 
-- (BOOL)canBecomeMainWindow
-{
+- (BOOL)canBecomeMainWindow {
     return YES;
 }
